refactor(listener): extract model toolset commands
