+++ src/websocket/listener/commands/types.ts
@@ -1,4 +1,5 @@
 import type WebSocket from "ws";
+import type { ConversationRuntime, ListenerRuntime } from "../types";
 
 export type SafeSocketSend = (
   socket: WebSocket,
@@ -11,3 +12,9 @@ export type RunDetachedListenerTask = (
   commandName: string,
   task: () => Promise<void>,
 ) => void;
+
+export type GetOrCreateScopedRuntime = (
+  listener: ListenerRuntime,
+  agentId?: string | null,
+  conversationId?: string | null,
+) => ConversationRuntime;
